Added (load) to Standard.
(load) is capable  of loading both .yst files  (files written with Raw
syntax and Standard initial scope) *and* Haskell modules containing

ystModule :: IO YLM.Data.Scope

There is  one caveat, however.   If the type  of ystModule is  not `IO
Scope`, but it is defined, no  error will be raised. Instead, some odd
corruption and/or crashes will occur. I'm working on a fix for this.

The module  loading capability  will probably be  added to  YLM proper
soon. It's quite big, and it's also something that makes sense to have
in the core.

The I/O  functions of Standard are  going to be moved  into a separate
(load)-able  module once Standard's  package system  is in  place. The
code to use them will probably be something like this:

    (using io)
    (io:put-line '"hello, how are you?")

Since I'm talking about the package system anyway, here's what you can
do with (using):

    (using io as input-output-system) ; imports io and namespaces it as input-output-system.
                                      ; that is, instead of calling io:put-line, you'd be calling
                                      ; input-output-system:put-line. useless here, but very useful
                                      ; for long package names.

    (using io *) ; imports io without any namespacing. instead of
                 ; calling io:put-line, you'd be calling just plain
                 ; ol' (put-line).

    (using io (put-line)) ; imports ONLY put-line from io, and puts
                          ; it in scope without any namespacing.